+++ src/vscode/extension.ts
@@ -43,6 +43,8 @@ class FactorioModConfigurationProvider implements vscode.DebugConfigurationProvi
 		const activeVersion = await this.versionSelector.getActiveVersion();
 		if (!activeVersion) { return; }
 
+		await activeVersion.checkSteamAppID(vscode.window);
+
 		if (await activeVersion.isPrototypeCacheEnabled()) {
 			const pcache = await vscode.window.showWarningMessage(
 				"Prototype Caching is enabled, which usually conflicts with the final portion of debugger initialization (which occurs in settings stage).",