+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * VSCode:
   * Migrate Mod Portal API key from VSCode key storage to `keytar` for consistent behavior across packages
   * Custom Editor for `mod-settings.dat`
+  * Re-check if active version is Steam before launching debug
 
 ## 1.1.28
 * Debugger: